backend/requirements.txt
flask
flask-cors
tensorflow
pillow
numpy
python-dotenv
google-generativeai
requests

